refactor(app): split builder.rs into routing/openapi/health/run
Closes the primary #201 follow-up: builder.rs reduced from ~61KB to ~12KB.

- routing.rs: route registration, nest, static files, compression

- openapi.rs: schema, docs, swagger-ui

- health.rs: health endpoints, production_defaults

- run.rs: server run, apply helpers, http3

225 lib tests pass; public API snapshots unchanged.
